added question.py and the data file
added question.py so it would be easier to keep all the questions in one place. added the data file to hold all the information on the player data so the can continue where the left off. got the creatures set up in a dictionary so it would be easier to make adjustments to them.